Added api subcommad to run the 3 new commands, moved api handling to new file, updated parser tests.
