+++ src/cli/mod.rs
@@ -3,4 +3,4 @@ mod args;
 mod headers;
 mod version;
 
-pub use args::{Args, Command, EvaluationCommand, ProviderCommand};
+pub use args::{ApiCommand, Args, Command, EvaluationCommand, ProviderCommand};
